+++ NEWS.md
@@ -1,6 +1,17 @@
+
+# reporter 1.2.0
+
+* Added Arial, Times, and Courier fonts to RTF output type.  This change required a 
+significant rewrite to RTF report creation functions.  Font is controlled using
+the _font_ parameter on the `create_report()` function.  Old text-style RTF
+output is still available using the "fixed" font.
+
 # reporter 1.1.4
 
 * Added support for covr and codecov.
+* Fixed overflow bug on RTF on Windows when using 12pt font.
+* Fixed line length/alignment issues on titles and footnotes.
+* Added _valign_ parameter to `footnotes()` function.
 
 # reporter 1.1.3
 
@@ -19,7 +30,6 @@ Curly brace escape will remain active for backward compatibility, but is
 no longer documented.
 * Fixed bug in RTF that was causing page overflows in LibreOffice Writer.
 * Various other bug fixes.
-
 
 # reporter 1.1.2